Refactor Ollama provider helpers into package
