+++ projects/04-llm-adapter-shadow/src/llm_adapter/providers/ollama/__init__.py
@@ -1,0 +1,5 @@
+"""Ollama provider package."""
+
+from .provider import DEFAULT_HOST, OllamaProvider
+
+__all__ = ["OllamaProvider", "DEFAULT_HOST"]
