+++ projects/04-llm-adapter-shadow/tests/providers/test_ollama_provider.py
@@ -4,7 +4,7 @@ import pytest
 
 from src.llm_adapter.errors import AuthError, RateLimitError, TimeoutError
 from src.llm_adapter.provider_spi import ProviderRequest
-from src.llm_adapter.providers.ollama import OllamaProvider
+from src.llm_adapter.providers.ollama.provider import OllamaProvider
 
 from tests.helpers.fakes import FakeResponse, FakeSession
